+++ backend/cmd/backend/main.go
@@ -1,0 +1,14 @@
+// @title Restaurateur API
+// @version 0.2.0
+// @description Provides info about restaurants in Prague
+// @host localhost:8080
+// @BasePath /
+package main
+
+import (
+	"github.com/AgiliaErnis/restaurateur/backend/pkg/api"
+)
+
+func main() {
+	api.Run()
+}
