+++ backend/pkg/api/docs/swagger.json
@@ -1,29 +1,14 @@
-// GENERATED BY THE COMMAND ABOVE; DO NOT EDIT
-// This file was generated by swaggo/swag
-
-package docs
-
-import (
-	"bytes"
-	"encoding/json"
-	"strings"
-
-	"github.com/alecthomas/template"
-	"github.com/swaggo/swag"
-)
-
-var doc = `{
-    "schemes": {{ marshal .Schemes }},
+{
     "swagger": "2.0",
     "info": {
-        "description": "{{.Description}}",
-        "title": "{{.Title}}",
+        "description": "Provides info about restaurants in Prague",
+        "title": "Restaurateur API",
         "contact": {},
         "license": {},
-        "version": "{{.Version}}"
+        "version": "0.2.0"
     },
-    "host": "{{.Host}}",
-    "basePath": "{{.BasePath}}",
+    "host": "localhost:8080",
+    "basePath": "/",
     "paths": {
         "/autocomplete": {
             "get": {
@@ -387,118 +372,6 @@ var doc = `{
                     }
                 }
             }
-        },
-        "/user": {
-            "get": {
-                "description": "Returns a JSON with user info if the request headers contain an authenticated cookie.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "user"
-                ],
-                "summary": "Get info about a user",
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/api.responseUserJSON"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/api.responseSimpleJSON"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "type": "string"
-                        }
-                    }
-                }
-            },
-            "delete": {
-                "description": "Deletes a user if the request headers contain an authenticated cookie.",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "Delete user"
-                ],
-                "summary": "Deletes a user",
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/api.responseSimpleJSON"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/api.responseSimpleJSON"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "type": "string"
-                        }
-                    }
-                }
-            },
-            "patch": {
-                "description": "Updates user's password or username based on the provided JSON. Only 1 field can be updated at a time. For password you need to provide \"oldPassword\" and \"newPassword\" fields, omitting the \"newUsername\" field and vice versa if you'd like to update the username",
-                "consumes": [
-                    "application/json"
-                ],
-                "produces": [
-                    "application/json"
-                ],
-                "tags": [
-                    "Patch user"
-                ],
-                "summary": "Updates a user's password or username",
-                "parameters": [
-                    {
-                        "description": "Create a new user",
-                        "name": "updateJSON",
-                        "in": "body",
-                        "required": true,
-                        "schema": {
-                            "$ref": "#/definitions/api.userUpdate"
-                        }
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "OK",
-                        "schema": {
-                            "$ref": "#/definitions/api.responseSimpleJSON"
-                        }
-                    },
-                    "400": {
-                        "description": "Bad Request",
-                        "schema": {
-                            "$ref": "#/definitions/api.responseSimpleJSON"
-                        }
-                    },
-                    "500": {
-                        "description": "Internal Server Error",
-                        "schema": {
-                            "type": "string"
-                        }
-                    }
-                }
-            }
         }
     },
     "definitions": {
@@ -605,25 +478,6 @@ var doc = `{
                 }
             }
         },
-        "api.userUpdate": {
-            "type": "object",
-            "required": [
-                "newPassword",
-                "newUsername",
-                "oldPassword"
-            ],
-            "properties": {
-                "newPassword": {
-                    "type": "string"
-                },
-                "newUsername": {
-                    "type": "string"
-                },
-                "oldPassword": {
-                    "type": "string"
-                }
-            }
-        },
         "db.User": {
             "type": "object",
             "required": [
@@ -644,51 +498,4 @@ var doc = `{
             }
         }
     }
-}`
-
-type swaggerInfo struct {
-	Version     string
-	Host        string
-	BasePath    string
-	Schemes     []string
-	Title       string
-	Description string
-}
-
-// SwaggerInfo holds exported Swagger Info so clients can modify it
-var SwaggerInfo = swaggerInfo{
-	Version:     "0.2.0",
-	Host:        "localhost:8080",
-	BasePath:    "/",
-	Schemes:     []string{},
-	Title:       "Restaurateur API",
-	Description: "Provides info about restaurants in Prague",
-}
-
-type s struct{}
-
-func (s *s) ReadDoc() string {
-	sInfo := SwaggerInfo
-	sInfo.Description = strings.Replace(sInfo.Description, "\n", "\\n", -1)
-
-	t, err := template.New("swagger_info").Funcs(template.FuncMap{
-		"marshal": func(v interface{}) string {
-			a, _ := json.Marshal(v)
-			return string(a)
-		},
-	}).Parse(doc)
-	if err != nil {
-		return doc
-	}
-
-	var tpl bytes.Buffer
-	if err := t.Execute(&tpl, sInfo); err != nil {
-		return doc
-	}
-
-	return tpl.String()
-}
-
-func init() {
-	swag.Register(swag.Name, &s{})
 }
